Populate patients-grid with all necessary data from server-side. (#36)
Before this only the names of patients were sent to the client. Now every column of the grid is populated with data that is converted into supported types.

Couldn't access backend from the test-case so patient's name and id are hard-coded and for other values only the correct format is tested because the values change between server-sessions.
This commit also creates separate classes for test-cases and introduces a common superclass for them.