test: limit number of tabs to open for memory test
Opening a new tab for every UI to measure VaadinSession memory size makes the
tests fail because the browser may crash if too many tabs are opened or
because the expected memory usage is based on a small number of UI
concurrently opened.
This change limit the number of tabs to 5, so that the tests runs in
condition similar to previous Flow versions.