test: prevent login redirect for memory test
PatientsList memory test requires unauthenticated access to the view.
The workaround to bypass authentication was to set an attribute in Vaadin session.
However, after latest changes in Flow there two issues arise:
1. MemoryMeasurement sets a string value for the login session attribute that seems
to cause subsequent issues with the LoginView that expect it as a boolean.
2. PatientsListMemory parent layout beforeEnter method is called before
   the attribute gets set and causes a redirect to the login view, but
   then the attribute is set and login view redirect to another page,
   that is not the one expected by the test.

This change fixes the authentication attribute type and updates parent
layout beforeEnter logic to check for anonymous allowed annotation on
target view before redirecting to login.

workaround to prevent redirect to login for memory test